requirements.txt
flask
flask-cors
pandas
google-cloud-bigquery
elasticsearch
tmdbv3api
requests
db-dtypes